mcp_server/requirements.txt
fastapi>=0.100.0
uvicorn
httpx
supabase
google-generativeai
python-dotenv
pydantic>=2.0.0
pydantic-settings
